+++ twilio.js
@@ -1,45 +1,32 @@
 require('dotenv').config()
 const twilio = require('twilio')
-const express = require('express')
-const messagingResponse = twilio.twiml.MessagingResponse
+const bodyParser = require('body-parser')
+const MessagingResponse = twilio.twiml.MessagingResponse
 var client;
 
 
 module.exports = {
     config: function() {
+        // Twilio config
         client = new twilio(process.env.accountSid, process.env.authToken);
-        
         console.log("STATUS: Twilio setup complete.");
+    },
 
-        // Set up express on port 3000
-        const port = 3000
-        const app = express()
-
+    startListener: function(expressObject) {
+        expressObject.use(bodyParser.urlencoded({ extended: false }));
+        
         // SMS listener
-        app.post('/sms', (req, res) => {
-            console.log(req)
-            // const twiml = new MessagingResponse()
+        expressObject.post('/sms', (req, res) => {
 
-            messagingResponse.message("asdf")
+            console.log("Message: " + req.body.Body + " " + req.body.From)
+            const twiml = new MessagingResponse()
+
+            twiml.message("asdf")
             res.writeHead(200, {'Content-Type': 'text/xml'});
             res.end(twiml.toString());
         })
 
-        // Web GUI
-        app.post('/create-game', (req, res) => {
-            console.log(req)
-            let newGameID
-            while (runningGames[(newGameID = genUniqueID())] !== undefined) {}
-            runningGames[newGameID] = new Game()
-            res.json({ id: newGameID })
-        })
-
-        app.use(express.static('public'))
-        app.listen(port, () => console.log(`STATUS: Web server and SMS listener active on port ${port}!`))
-
-
-
-
+        console.log("Status: SMS listener enabled.")
     },
 
     sendMessage: function(text, toNumber) {
@@ -49,10 +36,5 @@ module.exports = {
             to: toNumber
         })
         .then((message) => console.log(message.sid));
-    },
-
-    receiveMessage: function() {
-        
     }
-
 };
